# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Player API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Thumbnail {
  url?: string;
  width?: number;
  height?: number;
}

export type TrackArtistsItem = {
  name?: string;
  id?: string;
};

export type TrackAlbum = {
  name?: string;
  id?: string;
} | null;

export interface Track {
  videoId: string;
  title: string;
  artists: TrackArtistsItem[];
  album?: TrackAlbum;
  duration?: string | null;
  durationSeconds?: number | null;
  thumbnails: Thumbnail[];
  isExplicit?: boolean | null;
}

export interface SearchResults {
  results: Track[];
  query: string;
}

export type SongInfoArtistsItem = {
  name?: string;
};

export interface SongInfo {
  videoId: string;
  title: string;
  artists: SongInfoArtistsItem[];
  album?: string | null;
  thumbnails: Thumbnail[];
  duration?: string | null;
  durationSeconds?: number | null;
  youtubeUrl: string;
}

export interface TopArtist {
  title: string;
  browseId: string;
  subscribers?: string | null;
  thumbnails: Thumbnail[];
  rank?: string | null;
  trend?: string | null;
}

export interface ChartResults {
  chartTracks: Track[];
  topArtists: TopArtist[];
  featuredTitle: string;
  country: string;
}

export interface SearchSuggestions {
  suggestions: string[];
}

export interface WatchPlaylist {
  tracks: Track[];
  playlistId?: string | null;
}

export interface AlbumDetail {
  albumId: string;
  title: string;
  artist: string;
  year?: string | null;
  thumbnails: Thumbnail[];
  tracks: Track[];
  trackCount: number;
  description?: string | null;
}

export type SearchMusicParams = {
  q: string;
  filter?: SearchMusicFilter;
};

export type SearchMusicFilter =
  (typeof SearchMusicFilter)[keyof typeof SearchMusicFilter];

export const SearchMusicFilter = {
  songs: "songs",
  videos: "videos",
  albums: "albums",
  artists: "artists",
  playlists: "playlists",
} as const;

export type GetChartsParams = {
  country?: string;
};

export type GetSearchSuggestionsParams = {
  q: string;
};