# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Music Player API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: music
    description: Music operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /music/search:
    get:
      operationId: searchMusic
      tags: [music]
      summary: Search for music
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: filter
          in: query
          required: false
          schema:
            type: string
            enum: [songs, videos, albums, artists, playlists]
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResults"

  /music/song/{videoId}:
    get:
      operationId: getSong
      tags: [music]
      summary: Get song info
      parameters:
        - name: videoId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Song info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SongInfo"

  /music/charts:
    get:
      operationId: getCharts
      tags: [music]
      summary: Get top charts
      parameters:
        - name: country
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Chart results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChartResults"

  /music/suggestions:
    get:
      operationId: getSearchSuggestions
      tags: [music]
      summary: Get search suggestions
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchSuggestions"

  /music/watch/{videoId}:
    get:
      operationId: getWatchPlaylist
      tags: [music]
      summary: Get watch playlist
      parameters:
        - name: videoId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Watch playlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchPlaylist"

  /music/album/{albumId}:
    get:
      operationId: getAlbum
      tags: [music]
      summary: Get album details and tracks
      parameters:
        - name: albumId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Album details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlbumDetail"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Thumbnail:
      type: object
      properties:
        url:
          type: string
        width:
          type: integer
        height:
          type: integer

    Track:
      type: object
      properties:
        videoId:
          type: string
        title:
          type: string
        artists:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              id:
                type: string
        album:
          type: object
          nullable: true
          properties:
            name:
              type: string
            id:
              type: string
        duration:
          type: string
          nullable: true
        durationSeconds:
          type: integer
          nullable: true
        thumbnails:
          type: array
          items:
            $ref: "#/components/schemas/Thumbnail"
        isExplicit:
          type: boolean
          nullable: true
      required:
        - videoId
        - title
        - artists
        - thumbnails

    SearchResults:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/Track"
        query:
          type: string
      required:
        - results
        - query

    SongInfo:
      type: object
      properties:
        videoId:
          type: string
        title:
          type: string
        artists:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
        album:
          type: string
          nullable: true
        thumbnails:
          type: array
          items:
            $ref: "#/components/schemas/Thumbnail"
        duration:
          type: string
          nullable: true
        durationSeconds:
          type: integer
          nullable: true
        youtubeUrl:
          type: string
      required:
        - videoId
        - title
        - artists
        - thumbnails
        - youtubeUrl

    TopArtist:
      type: object
      properties:
        title:
          type: string
        browseId:
          type: string
        subscribers:
          type: string
          nullable: true
        thumbnails:
          type: array
          items:
            $ref: "#/components/schemas/Thumbnail"
        rank:
          type: string
          nullable: true
        trend:
          type: string
          nullable: true
      required:
        - title
        - browseId
        - thumbnails

    ChartResults:
      type: object
      properties:
        chartTracks:
          type: array
          items:
            $ref: "#/components/schemas/Track"
        topArtists:
          type: array
          items:
            $ref: "#/components/schemas/TopArtist"
        featuredTitle:
          type: string
        country:
          type: string
      required:
        - chartTracks
        - topArtists
        - featuredTitle
        - country

    SearchSuggestions:
      type: object
      properties:
        suggestions:
          type: array
          items:
            type: string
      required:
        - suggestions

    WatchPlaylist:
      type: object
      properties:
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/Track"
        playlistId:
          type: string
          nullable: true
      required:
        - tracks

    AlbumDetail:
      type: object
      properties:
        albumId:
          type: string
        title:
          type: string
        artist:
          type: string
        year:
          type: string
          nullable: true
        thumbnails:
          type: array
          items:
            $ref: "#/components/schemas/Thumbnail"
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/Track"
        trackCount:
          type: integer
        description:
          type: string
          nullable: true
      required:
        - albumId
        - title
        - artist
        - thumbnails
        - tracks
        - trackCount