# lib/api-zod/src/generated/types/albumDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Player API
 * OpenAPI spec version: 0.1.0
 */
import type { Thumbnail } from "./thumbnail";
import type { Track } from "./track";

export interface AlbumDetail {
  albumId: string;
  title: string;
  artist: string;
  year?: string | null;
  thumbnails: Thumbnail[];
  tracks: Track[];
  trackCount: number;
  description?: string | null;
}