# lib/api-zod/src/generated/types/chartResults.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Player API
 * OpenAPI spec version: 0.1.0
 */
import type { TopArtist } from "./topArtist";
import type { Track } from "./track";

export interface ChartResults {
  chartTracks: Track[];
  topArtists: TopArtist[];
  featuredTitle: string;
  country: string;
}