# lib/api-zod/src/generated/types/songInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Player API
 * OpenAPI spec version: 0.1.0
 */
import type { SongInfoArtistsItem } from "./songInfoArtistsItem";
import type { Thumbnail } from "./thumbnail";

export interface SongInfo {
  videoId: string;
  title: string;
  artists: SongInfoArtistsItem[];
  album?: string | null;
  thumbnails: Thumbnail[];
  duration?: string | null;
  durationSeconds?: number | null;
  youtubeUrl: string;
}