# lib/api-zod/src/generated/types/thumbnail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Player API
 * OpenAPI spec version: 0.1.0
 */

export interface Thumbnail {
  url?: string;
  width?: number;
  height?: number;
}