# lib/api-zod/src/generated/types/topArtist.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Player API
 * OpenAPI spec version: 0.1.0
 */
import type { Thumbnail } from "./thumbnail";

export interface TopArtist {
  title: string;
  browseId: string;
  subscribers?: string | null;
  thumbnails: Thumbnail[];
  rank?: string | null;
  trend?: string | null;
}