# lib/api-zod/src/generated/types/track.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Player API
 * OpenAPI spec version: 0.1.0
 */
import type { Thumbnail } from "./thumbnail";
import type { TrackAlbum } from "./trackAlbum";
import type { TrackArtistsItem } from "./trackArtistsItem";

export interface Track {
  videoId: string;
  title: string;
  artists: TrackArtistsItem[];
  album?: TrackAlbum;
  duration?: string | null;
  durationSeconds?: number | null;
  thumbnails: Thumbnail[];
  isExplicit?: boolean | null;
}